# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "DEML BACKEND APP API", "version": "2.0.0", "description": "" },
  "paths": {
...
    "/api/v1/telemetry/endpoints": {
      "post": {
        "operationId": "telemetry_api_ingest_endpoint_telemetry",
        "summary": "Ingest Endpoint Telemetry",
        "parameters": [],
        "responses": { "200": { "description": "OK" } },
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/TelemetryPayload" } }
          },
          "required": true
        }
      }
    },
    "/api/v1/telemetry/cookie-consent": {
      "post": {
        "operationId": "telemetry_api_save_cookie_consent",
        "summary": "Save Cookie Consent",
        "parameters": [],
        "responses": { "200": { "description": "OK" } },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CookieConsentPayload" }
            }
          },
          "required": true
        }
      }
    },
    "/api/v1/telemetry/subscribe": {
      "post": {
        "operationId": "telemetry_api_subscribe_newsletter",
        "summary": "Subscribe Newsletter",
        "parameters": [],
        "responses": { "200": { "description": "OK" } },
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SubscribePayload" } }
          },
          "required": true
        }
      }
    },
    "/api/v1/telemetry/technology": {
      "post": {
        "operationId": "telemetry_api_ingest_technology_telemetry",
        "summary": "Ingest Technology Telemetry",
        "parameters": [],
        "responses": { "200": { "description": "OK" } },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TelemetryDualStreamPayload" }
            }
          },
          "required": true
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TelemetryPayload": {
        "properties": {
          "tenant_id": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tenant Id"
          },
          "url": { "title": "Url", "type": "string" },
          "status_code": { "title": "Status Code", "type": "integer" },
          "response_time_ms": { "title": "Response Time Ms", "type": "number" },
          "ip_address": { "title": "Ip Address", "type": "string" },
          "is_active": { "title": "Is Active", "type": "boolean" },
          "telemetry_context": {
            "anyOf": [{ "additionalProperties": true, "type": "object" }, { "type": "null" }],
            "title": "Telemetry Context"
          }
        },
        "required": ["url", "status_code", "response_time_ms", "ip_address", "is_active"],
        "title": "TelemetryPayload",
        "type": "object"
      },
      "CookieConsentPayload": {
        "properties": {
          "necessary": { "title": "Necessary", "type": "boolean" },
          "analytical": { "title": "Analytical", "type": "boolean" },
          "marketing": { "title": "Marketing", "type": "boolean" }
        },
        "required": ["necessary", "analytical", "marketing"],
        "title": "CookieConsentPayload",
        "type": "object"
      },
      "SubscribePayload": {
        "properties": {
          "email": { "title": "Email", "type": "string" },
          "consent": { "title": "Consent", "type": "boolean" }
        },
        "required": ["email", "consent"],
        "title": "SubscribePayload",
        "type": "object"
      },
      "TelemetryDualStreamPayload": {
        "properties": {
          "tenant_id": { "title": "Tenant Id", "type": "string" },
          "url": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Url" },
          "stream_type": { "title": "Stream Type", "type": "string" },
          "tech_name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tech Name"
          },
          "version": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Version" },
          "manifest_type": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Manifest Type"
          },
          "manifest_content": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Manifest Content"
          }
        },
        "required": ["tenant_id", "stream_type"],
        "title": "TelemetryDualStreamPayload",
        "type": "object"
      },
...
    },
    "securitySchemes": { "IntegrationAPIKeyAuth": { "type": "http", "scheme": "bearer" } }
  },
  "servers": []
}